Changed player movement functions to Vector2f instead of Vector2i

+++ src/Player.hpp
@@ -14,7 +14,7 @@ public:
 
 	sf::Vector2i GetPosition();
 
-	void Move(sf::Vector2i aDistance);
+	void Move(sf::Vector2f aDistance);
 
 private:
 	Animation* animation;
